Revert "Refactor/project name change"
