+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.chainsafe.weipay",
+    package = "com.reactnavigationdemo",
 )
 
 android_resource(
     name = "res",
-    package = "com.chainsafe.weipay",
+    package = "com.reactnavigationdemo",
     res = "src/main/res",
 )
