+++ android/app/build.gradle
@@ -98,7 +98,7 @@ android {
 
 
     defaultConfig {
-        applicationId "com.chainsafe.weipay"
+        applicationId "com.reactnavigationdemo"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
