+++ android/app/src/main/java/com/reactnavigationdemo/MainActivity.java
@@ -1,4 +1,4 @@
-package com.weipay;
+package com.reactnavigationdemo;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "WeiPay";
+        return "reactNavigationDemo";
     }
 }
