+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "com.chainsafe.weipay",
-  "displayName": "com.chainsafe.weipay"
+  "name": "reactNavigationDemo",
+  "displayName": "reactNavigationDemo"
 }
