+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "com.chainsafe.weipay",
+  "name": "reactNavigationDemo",
   "version": "0.0.1",
   "private": true,
   "scripts": {
